[AI] [Detector Agent] 탐지데이터 수신 및 파싱 구조 구현 진행

+++ Models/AI/AiDetectionBox.cs
@@ -1,8 +1,11 @@
 ﻿namespace OpenCvWpfTracking.Models.AI
 {
+    /// <summary>
+    /// [AI Detector] 객체 1개 [Bounding Box] 정보
+    /// </summary>
     public class AiDetectionBox
     {
-        public int ObjectId { get; set; }
+        public long ObjectId { get; set; }
 
         public int ClassIndex { get; set; }
